[cli] Move comments API types
